Adafruit_NeoMatrix 8x8 - 64 RGB LED Pixel Matrix
A compilation of diferent Voids to be controlled by joystick. just ad a
nex voyd or existing to joystickShield.is.
controll your Adafruit_NeoMatrix 8x8 - 64 RGB LED Pixel Matrix with an
arduino a youstick and a  5v powersupply ( works fantastic for hours
with an powerpack of 5000mah ).
adjust brightness with strip.setBrightness(**) and
matrix.setBrightness(**) and add,remove,change rainbow colors with
matrix.Color(),  to use or to disable comands/lines use // or remove //
this is a little messy code tha works, but can be improved a lot! have
fun :)